Using PCINT vector directly for speed. Needs 328
I have pushed an updated DCServoWithSerialTuning sketch which uses the
PCINT vector directly (to see if this is better than the attachInterrupt
framework). This code will only work on the Atmega328

I tested on an Uno clone. It seems stable to 30kHz step pulses, above
that there are some miscounts.

I had to do a bit of pin transposition, so the STEP is now on pin 4 and
DIRECTION on pin 2

I also made the optimisation of letting the interrupt service routine
update a signed long value which then gets copied to the double
precision float only when needed by PID. I changed targetPosition to
targetPositionDouble and added targetPositionLong.